cambios dashboard

+++ client/src/components/AdminComponents/AllEvents.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import {
   deleteEvent,
   editUser,
-  getActivities,
+  adminGetActivities,
 } from "../../Redux trad/actions.js";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, useNavigate } from "react-router-dom";
@@ -19,12 +19,12 @@ import swal from "sweetalert";
 function AllEvents() {
   const navigate = useNavigate();
   const dispatch = useDispatch();
-  const allActivities = useSelector((state) => state.activities);
+  const allActivities = useSelector((state) => state.allActivities);
   const userActu = useSelector((state) => state.user);
   const adminState = userActu.admin;
-
+  console.log(allActivities);
   useEffect(() => {
-    dispatch(getActivities());
+    dispatch(adminGetActivities(userActu.id));
   }, []);
   useEffect(() => {
     if (!adminState) {
@@ -32,16 +32,32 @@ function AllEvents() {
     }
   }, [adminState]);
 
-  const handleDelete = async (e, id, userEmail) => {
-    e.preventDefault();
-
-    if (
-      window.confirm(
-        "¿Estás seguro que quieres eliminar este evento? Si lo eliminas, no podrás deshacer esta acción."
-      ) === true
-    ) {
-      dispatch(deleteEvent(id));
-      swal("Evento eliminado correctamente.");
+  const handleDelete = async (id, deletedAt) => {
+    if (deletedAt === null) {
+      if (
+        window.confirm(
+          "¿Estás seguro que quieres eliminar este evento? Si lo eliminas, no podrás deshacer esta acción."
+        ) === true
+      ) {
+        dispatch(deleteEvent(id));
+        swal("Evento eliminado correctamente.");
+      }
+    } else {
+      swal("El evento ya se encuentra eliminado");
+    }
+  };
+  const handleViewReports = (id, reports, user) => {
+    navigate(`/admin/events/allReports/${id}`, { state: { reports, user } });
+  };
+  const handleViewReviews = (id, reviews, user) => {
+    navigate(`/admin/events/allReviews/${id}`, { state: { reviews, user } });
+  };
+  const handleViewDetail = (id, deleted) => {
+    console.log(deleted);
+    if (!deleted) {
+      navigate(`/home/detail/${id}`);
+    } else {
+      swal("El evento ya se encuentra eliminado");
     }
   };
 
@@ -64,9 +80,17 @@ function AllEvents() {
                   <th className="bg-blue-500  p-2">IMAGEN</th>
                   <th className="bg-blue-500  p-2">NOMBRE</th>
                   <th className="bg-blue-500  p-2">LUGAR</th>
-                  <th className="bg-blue-500  p-2">EVENTO REPORTADO?</th>
                   <th className="bg-blue-500  p-2" colSpan="2">
-                    OPCIONES
+                    REPORTES
+                  </th>
+                  <th className="bg-blue-500  p-2" colSpan="2">
+                    REVIEWS
+                  </th>
+                  <th className="bg-blue-500  p-2" colSpan="2">
+                    ACTIVO?
+                  </th>
+                  <th className="bg-blue-500  p-2" colSpan="2">
+                    VER DETALLE
                   </th>
                 </tr>
               </thead>
@@ -92,21 +116,46 @@ function AllEvents() {
                             : u.place}
                         </td>
 
-                        <td className="p-2">{u.admin ? "SÍ" : "NO"}</td>
+                        <td className="p-2">{u.reportEvent.length}</td>
+                        <td
+                          className={`p-2 ${
+                            u.reportEvent.length ? "cursor-pointer" : ""
+                          }`}
+                          title="Ver reportes del evento"
+                          onClick={() =>
+                            handleViewReports(u.id, u.reportEvent, u.name)
+                          }
+                        >
+                          {u.reportEvent.length ? <View /> : ""}
+                        </td>
 
+                        <td className="p-2">{u.reviewEvent.length}</td>
+                        <td
+                          className="p-2 cursor-pointer"
+                          title="Ver reviews del evento"
+                          onClick={() =>
+                            handleViewReviews(u.id, u.reviewEvent, u.name)
+                          }
+                        >
+                          {u.reviewEvent.length ? <View /> : ""}
+                        </td>
+                        <td className="p-2">{u.deletedAt ? "NO" : "SI"}</td>
                         <td>
-                          <button className="text-blue-500 hover:text-blue-700 focus:outline-none">
-                            {" "}
-                            <Link to={`/home/detail/${u.id}`}>
-                              <View />
-                            </Link>
-                          </button>
-
                           <button
-                            onClick={(e) => handleDelete(e, u.id, u.email)}
+                            title="eliminar evento"
+                            onClick={(e) => handleDelete(u.id, u.deletedAt)}
                             className="text-red-500 hover:text-red-700 focus:outline-none ml-2"
                           >
                             <Remove />
+                          </button>
+                        </td>
+                        <td>
+                          <button
+                            title="Ver detalle del evento"
+                            className="text-blue-500 hover:text-blue-700 focus:outline-none"
+                            onClick={() => handleViewDetail(u.id, u.deletedAt)}
+                          >
+                            <View />
                           </button>
                         </td>
                       </tr>
